+++ CHANGELOG.md
@@ -11,6 +11,8 @@ _Feat_
 - Add retrofit to project + migration of http to Retrofit
 - Add units tests and coverage above 80%
 - Add github actions
+- Add max_design packages
+- Add configuration service
 
 _Fix_